refactor: extract escaped-quote scanner into jsonEscapedCharScanner.kt
Address review on #1831: move the scanner into a functions-only utility
file per Kotlin lowerCamelCase convention and expand the KDoc with
repaired vs untouched examples for each adjacent delimiter ({, [, :, ,).
Add an array-delimiter case to the tests to back the new examples.
